cpu-o3: change instTypes parameter  to uint16_t in getReg methods
To be able to store more instTypeFlag flags than 8 i went from using uint8_t to using uint16_t

+++ src/cpu/o3/cpu.cc
@@ -965,7 +965,7 @@ CPU::setMiscReg(int misc_reg, RegVal val, ThreadID tid)
 }
 
 RegVal
-CPU::getReg(PhysRegIdPtr phys_reg, uint8_t instTypes, ThreadID tid)
+CPU::getReg(PhysRegIdPtr phys_reg, uint16_t instTypes, ThreadID tid)
 {
     switch (phys_reg->classValue()) {
       case IntRegClass:
@@ -991,7 +991,7 @@ CPU::getReg(PhysRegIdPtr phys_reg, uint8_t instTypes, ThreadID tid)
 }
 
 void
-CPU::getReg(PhysRegIdPtr phys_reg, void *val, uint8_t instTypes, ThreadID tid)
+CPU::getReg(PhysRegIdPtr phys_reg, void *val, uint16_t instTypes, ThreadID tid)
 {
     switch (phys_reg->classValue()) {
       case IntRegClass:
@@ -1089,7 +1089,7 @@ CPU::getArchReg(const RegId &reg, ThreadID tid)
 {
     const RegId flat = reg.flatten(*isa[tid]);
     PhysRegIdPtr phys_reg = commitRenameMap[tid].lookup(flat);
-    return regFile.getReg(phys_reg, (uint8_t)0);
+    return regFile.getReg(phys_reg, (uint16_t)0);
 }
 
 void
@@ -1097,7 +1097,7 @@ CPU::getArchReg(const RegId &reg, void *val, ThreadID tid)
 {
     const RegId flat = reg.flatten(*isa[tid]);
     PhysRegIdPtr phys_reg = commitRenameMap[tid].lookup(flat);
-    regFile.getReg(phys_reg, val, (uint8_t)0);
+    regFile.getReg(phys_reg, val, (uint16_t)0);
 }
 
 void *
